requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0
ddgs>=0.5.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
psycopg2-binary>=2.9.0
pgvector>=0.3.0
asyncpg>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
sse-starlette>=2.0.0
tenacity>=9.0.0
lxml>=5.0.0
pytest>=8.0.0
